cleaned up some lines and made main more readable
